Refactor: add better support for competition select
When a user decides to add a new catch and is active in multiple competitions,
he is first presented with a modal window containing all of the active
competitions he is competing in. It forces him to select a competition before
adding a catch because a catch cannot exist outside of a competition.